+++ package.json
@@ -1,18 +1,12 @@
 {
-  "name": "agda-check",
-  "version": "1.1.1",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Lorenzobattistela/agda-check.git"
-  },
-  "bugs": {
-    "url": "https://github.com/Lorenzobattistela/agda-check/issues"
-  },
-  "homepage": "https://github.com/Lorenzobattistela/agda-check#readme",
-  "description": "A command-line tool to check Agda files and goals",
-  "main": "agda-check.js",
+  "name": "agda-cli",
+  "version": "1.0.3",
+  "description": "A command-line tool to perform Agda commands like checking, running, and compiling",
+  "main": "agda-cli.js",
   "bin": {
-    "agda-check": "./agda-check.js"
+    "agda-cli": "./agda-cli.js",
+    "agda-js": "./agda-js.js",
+    "agda-compile": "./agda-compile.js"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -20,10 +14,20 @@
   "keywords": [
     "agda",
     "check",
-    "cli"
+    "cli",
+    "compile",
+    "javascript"
   ],
   "author": "Lorenzo Battistela",
   "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Lorenzobattistela/agda-cli.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Lorenzobattistela/agda-cli/issues"
+  },
+  "homepage": "https://github.com/Lorenzobattistela/agda-cli#readme",
   "dependencies": {
     "child_process": "^1.0.2",
     "fs": "^0.0.1-security",
